+++ src/main/java/com/t2m/g2nee/front/bookset/category/controller/AdminCategoryController.java
@@ -1,6 +1,5 @@
 package com.t2m.g2nee.front.bookset.category.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.t2m.g2nee.front.bookset.category.service.CategoryService;
 import com.t2m.g2nee.front.bookset.category.dto.request.CategorySaveDto;
 import org.springframework.stereotype.Controller;
@@ -103,7 +102,7 @@ public class AdminCategoryController {
         CategorySaveDto request = new CategorySaveDto(categoryName, categoryEngName, isActivated, ancestorCategoryId);
         categoryService.modifyCategory(categoryId, request);
         return "redirect:/admin/categories";
-    }
+    }//카테고리 반영속도가 느림
 
 
     /**
