+++ src/main/java/com/t2m/g2nee/front/utils/HttpHeadersUtil.java
@@ -4,11 +4,9 @@ import com.t2m.g2nee.front.token.util.JwtUtil;
 import java.util.List;
 import java.util.Objects;
 import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
+
 
 /**
  * Header 생성의 중복을 방지하기 위한 유틸 클래스
@@ -24,9 +22,17 @@ public class HttpHeadersUtil {
         httpHeaders.setContentType(MediaType.APPLICATION_JSON);
         httpHeaders.setAccept(List.of(MediaType.APPLICATION_JSON));
 
-        ServletRequestAttributes servletRequestAttributes =
-                (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
-        HttpServletRequest request = servletRequestAttributes.getRequest();
+        Cookie accessToken = CookieUtil.findCookie(JwtUtil.ACCESS_COOKIE);
+        if (Objects.nonNull(accessToken)) {
+            httpHeaders.add("Authorization", JwtUtil.TOKEN_TYPE + accessToken.getValue());
+        }
+        return httpHeaders;
+    }
+
+    public static HttpHeaders makeFileHttpHeaders() {
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setContentType(MediaType.MULTIPART_FORM_DATA);
+        httpHeaders.setAccept(List.of(MediaType.APPLICATION_JSON));
 
         Cookie accessToken = CookieUtil.findCookie(JwtUtil.ACCESS_COOKIE);
         if (Objects.nonNull(accessToken)) {
