+++ src/main/resources/templates/admin/pointPolicy/adminPointPolicyList.html
@@ -72,7 +72,7 @@
                 <a class="page-link" th:href="@{'?page=' + ${pointPolicies.currentPage + 1}}">></a>
             </li>
             <li>
-                <a class="page-link" th:href="@{'?page=' + ${pointPolicies.totalPage}}"></a>
+                <a class="page-link" th:href="@{'?page=' + ${pointPolicies.totalPage}}">≫</a>
             </li>
         </ul>
     </nav>
