+++ src/main/resources/templates/order/packageList.html
@@ -1,0 +1,82 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>g2nee 포장지 목록</title>
+</head>
+<body>
+<h2 class="text-center">포장지 목록</h2>
+
+<table class="table table-bordered table-hover">
+    <thead>
+    <tr>
+        <th></th>
+        <th>포장지 명</th>
+        <th>가격</th>
+        <th>수량</th>
+    </tr>
+    </thead>
+
+    <tbody th:each="package : ${packages.data}">
+    <tr>
+        <td>
+            <img class="w-100 h-100" th:src="${package.url}" alt="포장지 사진">
+        </td>
+
+        <td>
+            <span th:text="${package.name}"></span>
+        </td>
+        <td>
+            <span th:text="${package.price}"></span>
+        </td>
+        <td>
+
+            <form id="packageForm" th:action="${dd}"
+                  th:method="POST">
+                <div class="input-group quantity mr-3" style="width: 130px;">
+                    <div class="input-group-btn">
+                        <button class="btn btn-primary btn-minus">
+                            <i class="fa fa-minus"></i>
+                        </button>
+                    </div>
+                    <input type="text" id="packageCount" class="form-control bg-secondary text-center" value="1">
+                    <div class="input-group-btn">
+                        <button class="btn btn-primary btn-plus">
+                            <i class="fa fa-plus"></i>
+                        </button>
+                    </div>
+                </div>
+
+                <input type="submit" class="btn btn-primary" value="저장">
+            </form>
+
+        </td>
+    </tr>
+    </tbody>
+</table>
+
+<br/><br/>
+<nav aria-label="...">
+    <ul class="pagination d-flex justify-content-center">
+        <li>
+            <a class="page-link" th:href="@{'?page=1'}">≪</a>
+        </li>
+        <li th:class="${packages.currentPage == 1 ? 'page-item disabled' : 'page-item'}">
+            <a class="page-link" th:href="@{'?page=' + ${packages.currentPage - 1 }}"><</a>
+        </li>
+        <li th:each="pageButton : ${#numbers.sequence(packages.startPage,packages.endPage)}"
+            th:class="${packages.currentPage == pageButton ?   'page-item active' : 'page-item'}">
+            <a class="page-link" th:href="@{'?page=' + ${pageButton}}" th:text="${pageButton}"></a>
+        </li>
+        <li th:class="${packages.currentPage == packages.totalPage ? 'page-item disabled' : 'page-item'}">
+            <a class="page-link" th:href="@{'?page=' + ${packages.currentPage + 1}}">></a>
+        </li>
+        <li>
+            <a class="page-link" th:href="@{'?page=' + ${packages.totalPage}}">≫</a>
+        </li>
+    </ul>
+</nav>
+
+<script src="/static/js/main.js"></script>
+</body>
+</html>
